bugfix: All sessions for a given controller were getting closed mistakenly
Even if there was error on one session for a given controller id (i.e. a given
iCloud account), all sessions for that controller id were getting closed.
Eg. If a user is logged in via an iPhone as well as Apple TV, if the iPhone
leaves the network, eventually causing its session to break, even the Apple TV
session breaks.

Some function names were changed to prevent confusion.

Closes: https://github.com/espressif/esp-homekit-sdk/issues/18