TCI048: Added a workaround to help pass TCI048
When the accessory state changes from unpaired to paired (after pair setup),
there is a Bonjour re-announcement required since the sf value changes.
During this re-annoucement, the accessory is in "disconnected" state, since
"connection" typically means a pair verified session, which happens after this
re-announcement. So, the s# value should also increment as per the specifications.

However, the HomeKit Certification Test Cases R10 require a different behaviour,
wherein the s# should not increment at this point. This commit adds a workaround
just for this test to pass.